Split out `CmdRepl` and `editorFor`
The REPL itself and the `nix repl` CLI are conceptually different
things, and thus deserve to be in different files.
